refactor: split lib into handle
